pyproject.toml
[tool.poetry]
name = "V-Bot"
version = "3.0.0"
description = ""
authors = ["Vioshim <Krangelj@unicartagena.edu.co>"]

[tool.poetry.dependencies]
python = "3.10.4"
pyright = "^0.0.13"
beautifulsoup4 = "^4.10.0"
pyaml = "^21.10.1"
yapf = "^0.31.0"
pytz = "^2021.3"
typing-extensions = "^4.0.1"
dateparser = "^1.1.0"
isort = {version = "^5.10.1", extras = ["requirements_deprecated_finder", "pipfile_deprecated_finder"]}
flake8 = "^4.0.1"
frozendict = "^2.1.3"
asyncpg = "^0.25.0"
nested-lookup = "^0.2.23"
"mystbin.py" = "^2.2.0"
asyncdagpi = "^4.0.0"
python-dotenv = "^0.19.2"
aiofiles = "^0.8.0"
uvloop = {version = "^0.16.0", platform = "linux"}
d20 = "^1.1.2"
wheel = "^0.37.1"
google-api-python-client = "^2.33.0"
google-auth-httplib2 = "^0.1.0"
google-auth-oauthlib = "^0.4.6"
aiohttp = "^3.8.1"
apscheduler = {git = "https://github.com/agronholm/apscheduler.git"}
jishaku = "^2.3.2"
python-docx = "^0.8.11"
colour = "^0.1.5"
humanize = "^4.0.0"
"discord.py" = {git = "https://github.com/Rapptz/discord.py.git", extras = ["speed", "voice"]}
rapidfuzz = "^2.0.10"
asyncpg-stubs = {git = "https://github.com/bryanforbes/asyncpg-stubs"}
aiogoogle = "^4.1.0"
motor = {version = "^3.0.0", extras = ["srv"]}
motor-stubs = "^1.7.1"
pymongo = {version = "^4.1.1", extras = ["srv"]}

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = {version = "^22.3.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]

line-length = 120 # override black's default line-length
target-version = ["py310"]
exclude = '''
/(
    \.git
  | \.mypy_cache
  | \.tox
  | venv
  | .vscode
  | .history
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''
[tool.pyright]
useLibraryCodeForTypes = true
typeCheckingMethod = "basic"
pythonVersion = "3.10.4"


[tool.isort]
# make it compatible with black
profile = "black" 
